Extract SSHProcessManager from terminal.py into ssh_process_manager.py
SSHProcessManager (a GTK-free singleton tracking spawned SSH processes and
terminating them on shutdown) moves verbatim out of the 6k-line, VTE-heavy
terminal.py into its own module. terminal.py re-exports SSHProcessManager and
the process_manager singleton, so existing 'from .terminal import ...' callers
(shutdown.py, main.py) need no change; the singleton is shared (same instance).

Pure structural move, no logic change. The only observable difference: the
process-manager log lines now originate from the 'ssh_process_manager' logger
instead of 'terminal' (same messages).

terminal.py: 6029 -> 5874 lines. Added characterization tests pinning the
singleton, terminal registration, process-group termination + SIGKILL
escalation, orphan skipping, and cleanup_all. Suite: 1645 -> 1653 passed.
